Long scoreboardId

+++ src/main/java/cn/nukkit/network/protocol/types/ScoreboardIdentityPacketEntry.java
@@ -2,7 +2,7 @@ package cn.nukkit.network.protocol.types;
 
 public class ScoreboardIdentityPacketEntry {
 
-    public int scoreboardId;
+    public long scoreboardId;
     public long entityUniqueId;
 
 }
